refactor: Stream 사용
